# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Técnico - UPD8",
        "description": "Desafio Técnico - UPD8",
        "contact": {
            "email": "acarlos.os@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://desafioupd8.teste/api",
            "description": "local"
        }
    ],
    "paths": {
...
        "/cliente": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Return clientes",
                "description": "Return clientes",
                "operationId": "cliente-index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "types": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "1"
                                                            },
                                                            "cidade_id": {
                                                                "type": "integer",
                                                                "example": "2"
                                                            },
                                                            "cpf": {
                                                                "type": "string",
                                                                "example": "99999999954"
                                                            },
                                                            "nome": {
                                                                "type": "string",
                                                                "example": "Pedro Oliveira"
                                                            },
                                                            "data_nascimento": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "1982-07-03T00:00:00.000000Z"
                                                            },
                                                            "sexo": {
                                                                "type": "enum",
                                                                "example": "masculino"
                                                            },
                                                            "endereco": {
                                                                "type": "string",
                                                                "example": "Rua Alves cardoso 144"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2024-07-09T22:22:07.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2024-07-09T22:22:07.000000Z"
                                                            },
                                                            "deleted_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "null"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Wrong error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong error. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Create cliente",
                "description": "Create cliente",
                "operationId": "cliente-store",
                "requestBody": {
                    "description": "Create cliente",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf",
                                    "data_nascimento",
                                    "sexo",
                                    "endereco",
                                    "cidade_id"
                                ],
                                "properties": {
                                    "cidade_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Maria Eduarda Oliveira"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "99999999912"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1982/07/03"
                                    },
                                    "sexo": {
                                        "type": "enum",
                                        "example": "masculino ou feminino"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Rua Alves cardoso 144"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "cidade_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Maria Eduarda Oliveira"
                                        },
                                        "cpf": {
                                            "type": "string",
                                            "example": "99999999912"
                                        },
                                        "data_nascimento": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1982/07/03"
                                        },
                                        "sexo": {
                                            "type": "enum",
                                            "example": "masculino ou feminino"
                                        },
                                        "endereco": {
                                            "type": "string",
                                            "example": "Rua Alves cardoso 144"
                                        },
                                        "uf": {
                                            "type": "string",
                                            "example": "SP"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-07-09T22:22:07.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-07-09T22:22:07.000000Z"
                                        },
                                        "deleted_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Wrong error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong error. Please try again!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}